Raise explicit exception when deleting config which has been referenced
This change aims to raise an explicit exception when trying to delete
a software config which has been referenced by deployment.

Closes-Bug: #1509915

+++ heat/db/sqlalchemy/api.py
@@ -915,6 +915,13 @@ def software_config_get_all(context, limit=None, marker=None,
 
 def software_config_delete(context, config_id):
     config = software_config_get(context, config_id)
+    # Query if the software config has been referenced by deployment.
+    result = model_query(context, models.SoftwareDeployment).filter_by(
+        config_id=config_id).first()
+    if result:
+        msg = (_("Software config with id %s can not be deleted as "
+                 "it is referenced.") % config_id)
+        raise exception.InvalidRestrictedAction(message=msg)
     session = orm_session.Session.object_session(config)
     with session.begin():
         session.delete(config)
